refactor: separate seed objects in files
