+++ prisma/seeds/addresses.ts
@@ -1,0 +1,10 @@
+export const addresses = [
+  {
+    address: 'ecash:qrmm7edwuj4jf7tnvygjyztyy0a0qxvl7quss2vxek',
+    networkId: 1
+  },
+  {
+    address: 'bitcoincash:qzqh7ej2vz26a9xaxq7capzfwgxt5gem9g8rvfxc5t',
+    networkId: 2
+  }
+]
