+++ prisma/seeds/connectors.ts
@@ -1,0 +1,18 @@
+export const connectors = [
+  {
+    addressId: 1,
+    paybuttonId: 1
+  },
+  {
+    addressId: 2,
+    paybuttonId: 2
+  },
+  {
+    addressId: 2,
+    paybuttonId: 3
+  },
+  {
+    addressId: 1,
+    paybuttonId: 3
+  }
+]
