+++ prisma/seeds/paybuttons.ts
@@ -15,33 +15,3 @@ export const paybuttons = [
     buttonData: '{}'
   }
 ]
-
-export const addresses = [
-  {
-    address: 'ecash:qrmm7edwuj4jf7tnvygjyztyy0a0qxvl7quss2vxek',
-    networkId: 1
-  },
-  {
-    address: 'bitcoincash:qzqh7ej2vz26a9xaxq7capzfwgxt5gem9g8rvfxc5t',
-    networkId: 2
-  }
-]
-
-export const connectors = [
-  {
-    addressId: 1,
-    paybuttonId: 1
-  },
-  {
-    addressId: 2,
-    paybuttonId: 2
-  },
-  {
-    addressId: 2,
-    paybuttonId: 3
-  },
-  {
-    addressId: 1,
-    paybuttonId: 3
-  }
-]
